+++ src/main/java/net/spell_engine/mixin/client/EntityRenderMixin.java
@@ -3,6 +3,7 @@ package net.spell_engine.mixin.client;
 import net.minecraft.client.render.Frustum;
 import net.minecraft.client.render.entity.EntityRenderer;
 import net.minecraft.entity.Entity;
+import net.spell_engine.internals.SpellHelper;
 import net.spell_engine.internals.casting.SpellCasterEntity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -14,7 +15,8 @@ public class EntityRenderMixin {
     @Inject(method = "shouldRender", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/Entity;getVisibilityBoundingBox()Lnet/minecraft/util/math/Box;"), cancellable = true)
     public void shouldRender_WhileBeaming(Entity entity, Frustum frustum, double x, double y, double z, CallbackInfoReturnable<Boolean> cir) {
         if (entity instanceof SpellCasterEntity caster) {
-            if (caster.isBeaming()) {
+            var spell = caster.getCurrentSpell();
+            if (spell != null && SpellHelper.isChanneled(spell)) {
                 cir.setReturnValue(true);
                 cir.cancel();
             }